test(tui): fix dead_code warnings in tests/common helpers
Move mouse() to m5.rs and buffer_to_string() to snapshots.rs (each
has a single consumer). Add a self-test for empty_graph() in common/mod.rs,
matching the pattern already used by every other common helper; this makes
the fn reachable in every binary and silences the per-binary lint.

Closes #1128
